pyproject.toml
[project]
name = "movielens"
version = "0.1.0"
description = "Experimentation on the movielens dataset"
readme = "README.md"
authors = [
    { name = "AnthonyBeech", email = "abeech123@virginmedia.com" }
]
requires-python = ">=3.11"
dependencies = [
    "feast>=0.20.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "mlflow>=2.19.0",
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "pandera>=0.22.1",
    "prefect>=3.1.13",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
]

[project.scripts]
movielens = "movielens:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "matplotlib>=3.10.0",
    "movielens",
    "pytest>=8.3.4",
    "ruff>=0.9.1",
    "yamllint>=1.35.1",
]

[tool.uv.sources]
movielens = { workspace = true }
